llmcoking/voice_agent_backend/pyproject.toml
[project]
name = "voice-agent-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "numpy>=1.24",
  "soundfile>=0.12",
  "torch>=2.0",
  "websockets>=16.0",
  "opencc-python-reimplemented>=0.1.7",
]

[tool.uvicorn]
factory = false
